log and restore table info after loop

+++ atlas-geo/processing/process-helpers.sh
@@ -108,15 +108,18 @@ run_dir() {
           log "$table got deleted."
         fi
         if [ "$FREEZE_DATA" == 1 ]; then
-          log "Restoring data. (.env 'FREEZE_DATA=1')"
           psql -q -c "DROP TABLE \"${table}\";"
           psql -q -c "ALTER TABLE \"${table}_backup\" RENAME TO \"${table}\";"
-          mv -f $backedup_tables $processed_tables
         else
           psql -q -c "DROP TABLE \"${table}_backup\";"
         fi
       done
-      rm $backedup_tables
+      if [ "$FREEZE_DATA" == 1 ]; then
+        log "Restored data (.env 'FREEZE_DATA=1')"
+        mv -f $backedup_tables $processed_tables
+      else
+        rm $backedup_tables
+      fi
       log_end "Compute Diffs"
     fi
   fi
